update travis.sh to only do work on tagged commits

+++ travis.sh
@@ -1,5 +1,8 @@
 #!/bin/sh
 
+# Only build tags
+if [ -z "$TRAVIS_TAG" ]; then exit 0; fi
+
 # Register binfmt_misc hooks
 docker run --rm --privileged multiarch/qemu-user-static:register --reset
